requirements.txt
pandas
scikit-learn
gunicorn
ipykernel
markdown
pytest
awscli
boto3
mlflow
dagshub
python-dotenv
gdown
dvc
mlflow
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
scipy
Flask
Flask-Cors
pymongo
plotly

-e .